support fileid with 64bit

+++ include/sync/fsync.h
@@ -1,9 +1,6 @@
 /*
  * Sync APIs - Any problem pls let me known. 
  * E-mail: chengtao786@gmail.com
- * /tmp/sync/source: vmsync_init(), filename_fileid_blocksize
- * /tmp/sync/send: vmsync_send(), filename_fileid_blocksize_blockid
- * /tmp/sync/fail: vmsync_send() function fails
  */
 
 #ifndef __FSYNC_H__
@@ -19,11 +16,11 @@ enum {
 	ERROR_SYNC_INIT ,
 };
 
-extern int vmsync_init(const char * work_path , const char * source_file_full_name, int file_id, uint32_t block_size);
-extern void vmsync_fini(const char * source_file_full_name, int file_id);
-extern int vmsync_send(int fild_id, uint64_t offset , uint64_t len);
-extern int vmsync_lock(int fild_id, uint64_t offset , uint64_t len);
-extern void vmsync_unlock(int fild_id, uint64_t offset , uint64_t len);
+extern int vmsync_init(const char * work_path , const char * source_file_full_name, uint64_t file_id, uint32_t block_size);
+extern void vmsync_fini(const char * source_file_full_name, uint64_t file_id);
+extern int vmsync_send(uint64_t fild_id, uint64_t offset, uint64_t len);
+extern int vmsync_lock(uint64_t fild_id, uint64_t offset, uint64_t len);
+extern void vmsync_unlock(uint64_t fild_id, uint64_t offset, uint64_t len);
 
 #endif
