+++ src/daemon/slave/priv.h
@@ -13,7 +13,7 @@ struct config_file_t{
 	int		use;
         int             fd;
         char            filename[512];
-        unsigned int    id;
+        uint64_t	id;
 };
 
 extern struct pthread_attr_t * slave_thread_alloc(void);
@@ -26,12 +26,12 @@ extern int do_accept(int listenfd);
 extern void * do_child(void *arg);
 
 extern int config_init(char * config_file);
-extern struct config_file_t * config_cft_lookup(__u32 file_id);
+extern struct config_file_t * config_cft_lookup(uint64_t file_id);
 
 extern int msg_handler(struct pthread_attr_t * pattr ,
                 struct config_file_t * cft ,
                 __u32 msg_type ,
-                __u32 msg_fileid ,
+                uint64_t msg_fileid ,
                 __u32 msg_blockid ,
 		__u8 * msg_data , 
                 __u32 msg_datalen);
